Update tst_bench_qhash hash functions to use size_t
Pre-requisite for a fix for qHash. The Qt50String inherits from QString
and becomes ambiguous once it no longer goes through a catch-all
template function because qHash(QString, size_t) has a better match for
the second argument.

(cherry picked from commit 7fea2d34fb3e9d13deb584595d119c870389b912)

+++ tests/benchmarks/corelib/tools/qhash/main.h
@@ -35,7 +35,7 @@ struct Qt4String : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const Qt4String &);
+size_t qHash(const Qt4String &);
 QT_END_NAMESPACE
 
 struct Qt50String : QString
@@ -45,7 +45,7 @@ struct Qt50String : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const Qt50String &, uint seed = 0);
+size_t qHash(const Qt50String &, size_t seed = 0);
 QT_END_NAMESPACE
 
 
@@ -56,6 +56,6 @@ struct JavaString : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const JavaString &);
+size_t qHash(const JavaString &);
 QT_END_NAMESPACE
